Decoupled configurations from CommerceManager.
