+++ Decoupled_Commerce_Application/Commerce.Core/Commerce.Core.csproj
@@ -48,10 +48,12 @@
     <Compile Include="Classes\CommerceManager.cs" />
     <Compile Include="Classes\CustomerValidator.cs" />
     <Compile Include="Classes\Logger.cs" />
+    <Compile Include="Classes\ConfigurationProviderFactory.cs" />
     <Compile Include="Configurations\CommerceAppConfigurationSection.cs" />
     <Compile Include="Configurations\CommerceAppElement.cs" />
     <Compile Include="Configurations\CustomerNotifierElement.cs" />
     <Compile Include="Configurations\PaymentProcessorElement.cs" />
+    <Compile Include="Interfaces\IConfigurationProviderFactory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
