+++ Decoupled_Commerce_Application/Commerce.Core/Interfaces/IConfigurationProviderFactory.cs
@@ -1,0 +1,10 @@
+﻿using Commerce.Shared.Contracts;
+
+namespace Commerce.Core
+{
+    public interface IConfigurationProviderFactory
+    {
+        IPaymentProcessor GetPaymentProcessor();
+        ICustomerNotifier GetCustomerNotifier();
+    }
+}
